daemon: thin plugins cli shell
