+++ crates/daemon/src/plugins_cli.rs
@@ -5,23 +5,23 @@ use std::path::Path;
 use clap::{Args, Subcommand, ValueEnum};
 use semver::Version;
 use serde::{Deserialize, Serialize};
-use serde_json::Value;
 
 use crate::kernel::{
-    CURRENT_PLUGIN_HOST_API, CURRENT_PLUGIN_MANIFEST_API_VERSION, Capability, ExecutionRoute,
-    HarnessKind, PACKAGE_MANIFEST_FILE_NAME, PluginBridgeKind, PluginCompatibility, PluginManifest,
-    VerticalPackManifest, plugin_runtime_scaffold_defaults,
+    CURRENT_PLUGIN_HOST_API, CURRENT_PLUGIN_MANIFEST_API_VERSION, Capability,
+    PACKAGE_MANIFEST_FILE_NAME, PluginBridgeKind, PluginCompatibility, PluginManifest,
+    plugin_runtime_scaffold_defaults,
 };
 use crate::{
-    BridgeSupportSpec, CliResult, HumanApprovalMode, HumanApprovalSpec, JsonSchemaDescriptor,
-    MaterializedBridgeSupportDeltaArtifact, OperationSpec, PluginInventoryResult,
-    PluginPreflightBridgeProfileRecommendation, PluginPreflightProfile, PluginPreflightResult,
-    PluginScanSpec, ResolvedBridgeSupportSelection, RunnerSpec, SecurityProfileSignatureSpec,
-    SpecRunReport, default_plugin_inventory_limit, default_plugin_preflight_limit, execute_spec,
+    BridgeSupportSpec, CliResult, JsonSchemaDescriptor, MaterializedBridgeSupportDeltaArtifact,
+    PluginInventoryResult, PluginPreflightProfile, PluginPreflightResult, execute_spec,
     json_schema_descriptor, materialize_bridge_support_delta_artifact,
-    materialize_bridge_support_template, resolve_bridge_support_policy,
-    resolve_bridge_support_selection,
+    materialize_bridge_support_template,
 };
+
+#[path = "plugins_cli_governance.rs"]
+mod governance_support;
+#[path = "plugins_cli_render.rs"]
+mod render_support;
 
 pub const PLUGINS_COMMAND_SCHEMA_VERSION: u32 = 1;
 pub const PLUGINS_COMMAND_SCHEMA_SURFACE: &str = "plugin_governance";
@@ -39,6 +39,17 @@ fn plugins_command_schema(purpose: &str) -> JsonSchemaDescriptor {
 
     json_schema_descriptor(version, surface, purpose)
 }
+
+use self::governance_support::{
+    build_plugin_doctor_context, build_plugin_inventory_context, build_plugin_preflight_context,
+    decode_plugin_inventory_results, decode_preflight_bridge_profile_recommendation,
+    decode_preflight_results, decode_preflight_summary, load_bridge_profile_views,
+    write_bridge_support_delta_artifact, write_bridge_support_template,
+};
+use self::render_support::{
+    render_plugins_cli_text, summarize_filtered_actions, summarize_plugin_doctor_results,
+    summarize_plugin_inventory_results,
+};
 
 #[derive(Subcommand, Debug, Clone, PartialEq, Eq)]
 pub enum PluginsCommands {
@@ -1343,1278 +1354,6 @@ fn render_plugin_scaffold_readme(package_root: &str, plugin_id: &str, bridge_kin
     ]
     .join("\n")
 }
-
-fn render_plugins_cli_text(execution: &PluginsCommandExecution) -> String {
-    let (title, body) = match execution {
-        PluginsCommandExecution::Init(execution) => {
-            ("plugins init", render_plugins_init_text(execution))
-        }
-        PluginsCommandExecution::Inventory(execution) => (
-            "plugins inventory",
-            render_plugins_inventory_text(execution),
-        ),
-        PluginsCommandExecution::Doctor(execution) => {
-            ("plugins doctor", render_plugins_doctor_text(execution))
-        }
-        PluginsCommandExecution::BridgeProfiles(execution) => (
-            "bridge profiles",
-            render_plugins_bridge_profiles_text(execution),
-        ),
-        PluginsCommandExecution::BridgeTemplate(execution) => (
-            "bridge template",
-            render_plugins_bridge_template_text(execution),
-        ),
-        PluginsCommandExecution::Preflight(execution) => (
-            "plugins preflight",
-            render_plugins_preflight_text(execution),
-        ),
-        PluginsCommandExecution::Actions(execution) => {
-            ("operator actions", render_plugins_actions_text(execution))
-        }
-    };
-    wrap_plugins_surface_text(title, body)
-}
-
-fn wrap_plugins_surface_text(title: &str, body: String) -> String {
-    crate::render_operator_shell_surface_from_body(title, "operator plugins", body)
-}
-
-fn render_plugins_init_text(execution: &PluginsInitExecution) -> String {
-    let source_language = execution.source_language.as_deref().unwrap_or("-");
-    let mut lines = vec![format!(
-        "plugins init package_root={} plugin_id={} provider_id={} connector_name={}",
-        execution.package_root,
-        execution.plugin_id,
-        execution.provider_id,
-        execution.connector_name
-    )];
-    lines.push(format!(
-        "- bridge_kind={} source_language={} adapter_family={} entrypoint={}",
-        execution.bridge_kind, source_language, execution.adapter_family, execution.entrypoint
-    ));
-    lines.push(format!("- manifest={}", execution.manifest_path));
-    lines.push(format!("- readme={}", execution.readme_path));
-    lines.push(format!(
-        "- next_steps=loong plugins doctor --root \"{}\" --profile sdk-release",
-        execution.package_root
-    ));
-    lines.push(format!(
-        "- operator_actions=loong plugins actions --root \"{}\" --profile sdk-release",
-        execution.package_root
-    ));
-    lines.join("\n")
-}
-
-fn render_plugins_inventory_text(execution: &PluginsInventoryExecution) -> String {
-    let mut lines = vec![format!(
-        "plugins inventory query={} roots={} returned_plugins={} ready={} setup_incomplete={} blocked={} deferred={} loaded={}",
-        display_text_or_dash(Some(execution.query.as_str())),
-        execution.scan_roots.join(","),
-        execution.returned_results,
-        execution.summary.ready_plugins,
-        execution.summary.setup_incomplete_plugins,
-        execution.summary.blocked_plugins,
-        execution.summary.deferred_plugins,
-        execution.summary.loaded_plugins
-    )];
-    lines.push(format!(
-        "bridge_support source={} sha256={}",
-        display_text_or_dash(execution.bridge_support_source.as_deref()),
-        display_text_or_dash(execution.bridge_support_sha256.as_deref())
-    ));
-    lines.push(format!(
-        "bridge_support_delta source={} sha256={}",
-        display_text_or_dash(execution.bridge_support_delta_source.as_deref()),
-        display_text_or_dash(execution.bridge_support_delta_sha256.as_deref())
-    ));
-    lines.push(format!(
-        "ecosystem source_kind={} bridge={} language={} setup_surface={} activation_status={}",
-        format_rollup_map(&execution.summary.source_kind_distribution),
-        format_rollup_map(&execution.summary.bridge_kind_distribution),
-        format_rollup_map(&execution.summary.source_language_distribution),
-        format_rollup_map(&execution.summary.setup_surface_distribution),
-        format_rollup_map(&execution.summary.activation_status_distribution)
-    ));
-    for result in &execution.results {
-        let activation_status = inventory_result_status_label(result);
-        let setup_surface = inventory_result_setup_surface_label(result);
-        let source_language = result.source_language.as_deref().unwrap_or("-");
-        let manifest_path = display_text_or_dash(result.package_manifest_path.as_deref());
-        let setup_mode = display_text_or_dash(result.setup_mode.as_deref());
-        let host_api = result
-            .compatibility
-            .as_ref()
-            .and_then(|compatibility| compatibility.host_api.as_deref());
-        let host_version_req = result
-            .compatibility
-            .as_ref()
-            .and_then(|compatibility| compatibility.host_version_req.as_deref());
-        let required_env_vars = format_csv_or_dash(&result.setup_required_env_vars);
-        let required_config_keys = format_csv_or_dash(&result.setup_required_config_keys);
-        let runtime_health = result
-            .runtime_health
-            .as_ref()
-            .map(|health| health.status.as_str());
-        let attestation = result
-            .activation_attestation
-            .as_ref()
-            .map(|attestation| attestation.integrity.as_str());
-        lines.push(format!(
-            "- plugin={} provider={} status={} loaded={} deferred={} bridge={} language={} setup_surface={}",
-            result.plugin_id,
-            result.provider_id,
-            activation_status,
-            result.loaded,
-            result.deferred,
-            result.bridge_kind,
-            source_language,
-            setup_surface
-        ));
-        lines.push(format!(
-            "  manifest={} setup_mode={} required_env={} required_config={} host_api={} host_version_req={}",
-            manifest_path,
-            setup_mode,
-            required_env_vars,
-            required_config_keys,
-            display_text_or_dash(host_api),
-            display_text_or_dash(host_version_req)
-        ));
-        lines.push(format!(
-            "  source={} bootstrap_hint={} runtime_health={} attestation={} summary={}",
-            result.source_path,
-            display_text_or_dash(result.bootstrap_hint.as_deref()),
-            display_text_or_dash(runtime_health),
-            display_text_or_dash(attestation),
-            display_text_or_dash(result.summary.as_deref())
-        ));
-        if let Some(reason) = result.activation_reason.as_deref() {
-            lines.push(format!("  activation_reason={reason}"));
-        }
-    }
-    lines.join("\n")
-}
-
-fn render_plugins_doctor_text(execution: &PluginsDoctorExecution) -> String {
-    let preflight_summary = &execution.preflight_summary;
-    let mut lines = vec![format!(
-        "plugins doctor profile={} query={} roots={} matched_plugins={} returned_plugins={} passed={} warned={} blocked={}",
-        execution.profile,
-        display_text_or_dash(Some(execution.query.as_str())),
-        execution.scan_roots.join(","),
-        execution.summary.matched_plugins,
-        execution.returned_results,
-        execution.summary.passed_plugins,
-        execution.summary.warned_plugins,
-        execution.summary.blocked_plugins
-    )];
-    lines.push(format!(
-        "policy source={} version={} checksum={} sha256={}",
-        preflight_summary.policy_source,
-        display_text_or_dash(preflight_summary.policy_version.as_deref()),
-        preflight_summary.policy_checksum,
-        preflight_summary.policy_sha256
-    ));
-    lines.push(format!(
-        "bridge_support source={} sha256={}",
-        display_text_or_dash(execution.bridge_support_source.as_deref()),
-        display_text_or_dash(execution.bridge_support_sha256.as_deref())
-    ));
-    lines.push(format!(
-        "bridge_support_delta source={} sha256={}",
-        display_text_or_dash(execution.bridge_support_delta_source.as_deref()),
-        display_text_or_dash(execution.bridge_support_delta_sha256.as_deref())
-    ));
-    lines.push(format!(
-        "ecosystem bridge={} language={} setup_surface={} activation_status={}",
-        format_rollup_map(&execution.summary.bridge_kind_distribution),
-        format_rollup_map(&execution.summary.source_language_distribution),
-        format_rollup_map(&execution.summary.setup_surface_distribution),
-        format_rollup_map(&execution.summary.activation_status_distribution)
-    ));
-    lines.push(format!(
-        "doctor_attention activation_ready={} setup_incomplete={} deferred={} loaded={} attention={} remediation_counts={}",
-        execution.summary.activation_ready_plugins,
-        execution.summary.setup_incomplete_plugins,
-        execution.summary.deferred_plugins,
-        execution.summary.loaded_plugins,
-        execution.summary.packages_requiring_author_attention,
-        format_rollup_map(&execution.summary.remediation_counts)
-    ));
-    lines.push(format!(
-        "doctor_actions recommended={} operator_actions={} packages_with_operator_actions={} operator_plan_by_kind={}",
-        execution.summary.total_recommended_actions,
-        execution.summary.total_operator_actions,
-        execution.summary.packages_with_operator_actions,
-        format_rollup_map(&preflight_summary.operator_action_counts_by_kind)
-    ));
-    lines.extend(render_bridge_profile_fit_lines(preflight_summary));
-    for result in &execution.results {
-        lines.extend(render_plugin_doctor_result_lines(result));
-    }
-    lines.join("\n")
-}
-
-fn render_plugins_bridge_profiles_text(execution: &PluginsBridgeProfilesExecution) -> String {
-    let mut lines = vec![format!(
-        "plugins bridge-profiles returned_profiles={}",
-        execution.profiles.len()
-    )];
-    for profile in &execution.profiles {
-        lines.push(format!(
-            "- profile={} version={} source={} checksum={} sha256={}",
-            profile.profile_id,
-            profile.policy_version.as_deref().unwrap_or("-"),
-            profile.source,
-            profile.checksum,
-            profile.sha256
-        ));
-        lines.push(format!(
-            "  bridges={} compatibility={} shims={} execute_process_stdio={} execute_http_json={} enforce_supported={} enforce_execution_success={}",
-            format_csv_or_dash(&profile.supported_bridges),
-            format_csv_or_dash(&profile.supported_compatibility_modes),
-            format_csv_or_dash(&profile.supported_compatibility_shims),
-            profile.execute_process_stdio,
-            profile.execute_http_json,
-            profile.enforce_supported,
-            profile.enforce_execution_success
-        ));
-        for shim in &profile.shim_support_profiles {
-            lines.push(format!(
-                "  shim={} family={} version={} dialects={} bridges={} adapter_families={} languages={}",
-                shim.shim_id,
-                shim.shim_family,
-                display_text_or_dash(shim.version.as_deref()),
-                format_csv_or_dash(&shim.supported_dialects),
-                format_csv_or_dash(&shim.supported_bridges),
-                format_csv_or_dash(&shim.supported_adapter_families),
-                format_csv_or_dash(&shim.supported_source_languages)
-            ));
-        }
-    }
-    lines.join("\n")
-}
-
-fn render_plugins_bridge_template_text(execution: &PluginsBridgeTemplateExecution) -> String {
-    let mut lines = vec![format!(
-        "plugins bridge-template profile={} query={} roots={} matched_plugins={} template_kind={}",
-        execution.profile,
-        display_text_or_dash(Some(execution.query.as_str())),
-        execution.scan_roots.join(","),
-        execution.summary.matched_plugins,
-        execution.template_kind
-    )];
-    lines.push(format!(
-        "bridge_support source={} sha256={}",
-        display_text_or_dash(execution.bridge_support_source.as_deref()),
-        display_text_or_dash(execution.bridge_support_sha256.as_deref())
-    ));
-    lines.push(format!(
-        "bridge_support_delta source={} sha256={}",
-        display_text_or_dash(execution.bridge_support_delta_source.as_deref()),
-        display_text_or_dash(execution.bridge_support_delta_sha256.as_deref())
-    ));
-    lines.extend(render_bridge_profile_fit_lines(&execution.summary));
-    lines.push(format!(
-        "template profile={} source={} version={} checksum={} sha256={} output={}",
-        execution.template_profile_id,
-        execution.template_source,
-        display_text_or_dash(execution.template_policy_version.as_deref()),
-        execution.template_checksum,
-        execution.template_sha256,
-        display_text_or_dash(execution.output_path.as_deref())
-    ));
-    lines.push(format!(
-        "template_delta base_profile={} base_source={} base_version={} checksum={} sha256={} output={}",
-        execution.delta_artifact.base_profile_id,
-        execution.delta_artifact.base_source,
-        display_text_or_dash(execution.delta_artifact.base_policy_version.as_deref()),
-        execution.delta_artifact.checksum,
-        execution.delta_artifact.sha256,
-        display_text_or_dash(execution.delta_output_path.as_deref())
-    ));
-    lines.push(format!(
-        "template_delta_support bridges={} compatibility={} adapter_families={} shims={} shim_profiles={} unresolved={}",
-        format_csv_or_dash(&execution.delta_artifact.delta.supported_bridges),
-        format_csv_or_dash(&execution.delta_artifact.delta.supported_compatibility_modes),
-        format_csv_or_dash(&execution.delta_artifact.delta.supported_adapter_families),
-        format_csv_or_dash(&execution.delta_artifact.delta.supported_compatibility_shims),
-        format_bridge_shim_profile_delta_artifact(&execution.delta_artifact.delta.shim_profile_additions),
-        format_csv_or_dash(&execution.delta_artifact.delta.unresolved_blocking_reasons)
-    ));
-    lines.push(format!(
-        "template_support bridges={} compatibility={} shims={} execute_process_stdio={} execute_http_json={} enforce_supported={} enforce_execution_success={}",
-        execution
-            .template
-            .supported_bridges
-            .iter()
-            .map(|bridge| bridge.as_str().to_owned())
-            .collect::<Vec<_>>()
-            .join(","),
-        execution
-            .template
-            .supported_compatibility_modes
-            .iter()
-            .map(|mode| mode.as_str().to_owned())
-            .collect::<Vec<_>>()
-            .join(","),
-        execution
-            .template
-            .supported_compatibility_shims
-            .iter()
-            .map(|shim| format!("{}:{}", shim.shim_id, shim.family))
-            .collect::<Vec<_>>()
-            .join(","),
-        execution.template.execute_process_stdio,
-        execution.template.execute_http_json,
-        execution.template.enforce_supported,
-        execution.template.enforce_execution_success
-    ));
-    lines.join("\n")
-}
-
-fn render_plugins_preflight_text(execution: &PluginsPreflightExecution) -> String {
-    let mut lines = vec![format!(
-        "plugins preflight profile={} query={} roots={} matched_plugins={} returned_plugins={} passed={} warned={} blocked={}",
-        execution.profile,
-        display_text_or_dash(Some(execution.query.as_str())),
-        execution.scan_roots.join(","),
-        execution.summary.matched_plugins,
-        execution.summary.returned_plugins,
-        execution.summary.passed_plugins,
-        execution.summary.warned_plugins,
-        execution.summary.blocked_plugins
-    )];
-    lines.push(format!(
-        "policy source={} version={} checksum={} sha256={}",
-        execution.summary.policy_source,
-        execution.summary.policy_version.as_deref().unwrap_or("-"),
-        execution.summary.policy_checksum,
-        execution.summary.policy_sha256
-    ));
-    lines.push(format!(
-        "bridge_support source={} sha256={}",
-        display_text_or_dash(execution.bridge_support_source.as_deref()),
-        display_text_or_dash(execution.bridge_support_sha256.as_deref())
-    ));
-    lines.push(format!(
-        "bridge_support_delta source={} sha256={}",
-        display_text_or_dash(execution.bridge_support_delta_source.as_deref()),
-        display_text_or_dash(execution.bridge_support_delta_sha256.as_deref())
-    ));
-    lines.push(format!(
-        "ecosystem source_kind={} dialect={} compatibility={} language={} bridge={}",
-        format_rollup_map(&execution.summary.source_kind_distribution),
-        format_rollup_map(&execution.summary.dialect_distribution),
-        format_rollup_map(&execution.summary.compatibility_mode_distribution),
-        format_rollup_map(&execution.summary.source_language_distribution),
-        format_rollup_map(&execution.summary.bridge_kind_distribution)
-    ));
-    lines.push(format!(
-        "diagnostics total={} blocking={} error={} warning={} info={}",
-        execution.summary.total_diagnostics,
-        execution.summary.blocking_diagnostics,
-        execution.summary.error_diagnostics,
-        execution.summary.warning_diagnostics,
-        execution.summary.info_diagnostics
-    ));
-    lines.push(format!(
-        "operator_actions total={} by_surface={} by_kind={} reload={} no_reload={}",
-        execution.summary.operator_action_plan.len(),
-        format_rollup_map(&execution.summary.operator_action_counts_by_surface),
-        format_rollup_map(&execution.summary.operator_action_counts_by_kind),
-        execution.summary.operator_actions_requiring_reload,
-        execution.summary.operator_actions_without_reload
-    ));
-    lines.extend(render_bridge_profile_fit_lines(&execution.summary));
-    for result in &execution.results {
-        let plugin = &result.plugin;
-        let action_kinds =
-            format_preflight_result_operator_action_kinds(&result.recommended_actions);
-        lines.push(format!(
-            "- plugin={} provider={} verdict={} baseline={} activation_ready={} loaded={} actions={}",
-            plugin.plugin_id,
-            plugin.provider_id,
-            result.verdict,
-            result.baseline_verdict,
-            result.activation_ready,
-            plugin.loaded,
-            action_kinds
-        ));
-    }
-    lines.join("\n")
-}
-
-fn render_plugin_doctor_result_lines(result: &PluginPreflightResult) -> Vec<String> {
-    let plugin = &result.plugin;
-    let activation_status = inventory_result_status_label(plugin);
-    let setup_surface = inventory_result_setup_surface_label(plugin);
-    let source_language = plugin.source_language.as_deref().unwrap_or("-");
-    let manifest_path = display_text_or_dash(plugin.package_manifest_path.as_deref());
-    let setup_mode = display_text_or_dash(plugin.setup_mode.as_deref());
-    let required_env_vars = format_csv_or_dash(&plugin.setup_required_env_vars);
-    let required_config_keys = format_csv_or_dash(&plugin.setup_required_config_keys);
-    let setup_remediation = display_text_or_dash(plugin.setup_remediation.as_deref());
-    let runtime_health = plugin
-        .runtime_health
-        .as_ref()
-        .map(|health| health.status.as_str());
-    let attestation = plugin
-        .activation_attestation
-        .as_ref()
-        .map(|value| value.integrity.as_str());
-    let effective_flags = format_csv_or_dash(&result.effective_policy_flags);
-    let remediation_classes = format_preflight_remediation_classes(&result.remediation_classes);
-    let operator_action_kinds =
-        format_preflight_result_operator_action_kinds(&result.recommended_actions);
-    let blocking_diagnostics = format_csv_or_dash(&result.blocking_diagnostic_codes);
-    let advisory_diagnostics = format_csv_or_dash(&result.advisory_diagnostic_codes);
-    let recommended_actions =
-        format_preflight_result_recommended_actions(&result.recommended_actions);
-
-    let mut lines = vec![format!(
-        "- plugin={} provider={} verdict={} activation_status={} loaded={} deferred={} bridge={} language={} setup_surface={}",
-        plugin.plugin_id,
-        plugin.provider_id,
-        result.verdict,
-        activation_status,
-        plugin.loaded,
-        plugin.deferred,
-        plugin.bridge_kind,
-        source_language,
-        setup_surface
-    )];
-    lines.push(format!(
-        "  manifest={} setup_mode={} required_env={} required_config={} setup_remediation={}",
-        manifest_path, setup_mode, required_env_vars, required_config_keys, setup_remediation
-    ));
-    lines.push(format!(
-        "  source={} activation_ready={} runtime_health={} attestation={} summary={}",
-        plugin.source_path,
-        result.activation_ready,
-        display_text_or_dash(runtime_health),
-        display_text_or_dash(attestation),
-        display_text_or_dash(plugin.summary.as_deref())
-    ));
-    lines.push(format!(
-        "  policy_summary={} effective_flags={} remediation_classes={} operator_actions={}",
-        result.policy_summary, effective_flags, remediation_classes, operator_action_kinds
-    ));
-    lines.push(format!(
-        "  blocking_diagnostics={} advisory_diagnostics={}",
-        blocking_diagnostics, advisory_diagnostics
-    ));
-    if let Some(reason) = plugin.activation_reason.as_deref() {
-        lines.push(format!("  activation_reason={reason}"));
-    }
-    if recommended_actions != "-" {
-        lines.push(format!("  recommended_actions={recommended_actions}"));
-    }
-    lines
-}
-
-fn format_preflight_remediation_classes(
-    values: &[crate::PluginPreflightRemediationClass],
-) -> String {
-    if values.is_empty() {
-        return "-".to_owned();
-    }
-
-    let mut classes = values
-        .iter()
-        .map(|value| value.as_str().to_owned())
-        .collect::<Vec<_>>();
-    classes.sort();
-    classes.dedup();
-    classes.join(",")
-}
-
-fn format_preflight_result_operator_action_kinds(
-    values: &[crate::PluginPreflightRecommendedAction],
-) -> String {
-    let kinds = values
-        .iter()
-        .filter_map(|value| value.operator_action.as_ref())
-        .map(|value| value.kind.as_str().to_owned())
-        .collect::<BTreeSet<_>>();
-
-    if kinds.is_empty() {
-        return "-".to_owned();
-    }
-
-    kinds.into_iter().collect::<Vec<_>>().join(",")
-}
-
-fn format_preflight_result_recommended_actions(
-    values: &[crate::PluginPreflightRecommendedAction],
-) -> String {
-    if values.is_empty() {
-        return "-".to_owned();
-    }
-
-    let mut rendered = Vec::new();
-    for value in values {
-        let remediation_class = value.remediation_class.as_str();
-        let mut parts = vec![remediation_class.to_owned(), value.summary.clone()];
-        if let Some(action) = value.operator_action.as_ref() {
-            let kind = action.kind.as_str();
-            parts.push(format!("action={kind}"));
-        }
-        rendered.push(parts.join("|"));
-    }
-    rendered.join("; ")
-}
-
-fn render_plugins_actions_text(execution: &PluginsActionsExecution) -> String {
-    let mut lines = vec![format!(
-        "plugins actions profile={} query={} roots={} total_actions={} matched_actions={}",
-        execution.profile,
-        display_text_or_dash(Some(execution.query.as_str())),
-        execution.scan_roots.join(","),
-        execution.total_actions,
-        execution.matched_actions
-    )];
-    lines.push(format!(
-        "policy source={} version={} checksum={} sha256={}",
-        execution.summary.policy_source,
-        execution.summary.policy_version.as_deref().unwrap_or("-"),
-        execution.summary.policy_checksum,
-        execution.summary.policy_sha256
-    ));
-    lines.push(format!(
-        "bridge_support source={} sha256={}",
-        display_text_or_dash(execution.bridge_support_source.as_deref()),
-        display_text_or_dash(execution.bridge_support_sha256.as_deref())
-    ));
-    lines.push(format!(
-        "bridge_support_delta source={} sha256={}",
-        display_text_or_dash(execution.bridge_support_delta_source.as_deref()),
-        display_text_or_dash(execution.bridge_support_delta_sha256.as_deref())
-    ));
-    lines.push(format!(
-        "ecosystem source_kind={} dialect={} compatibility={} language={} bridge={}",
-        format_rollup_map(&execution.summary.source_kind_distribution),
-        format_rollup_map(&execution.summary.dialect_distribution),
-        format_rollup_map(&execution.summary.compatibility_mode_distribution),
-        format_rollup_map(&execution.summary.source_language_distribution),
-        format_rollup_map(&execution.summary.bridge_kind_distribution)
-    ));
-    lines.push(format!(
-        "filters surface={} kind={} requires_reload={}",
-        format_csv_or_dash(&execution.filters.surface),
-        format_csv_or_dash(&execution.filters.kind),
-        execution
-            .filters
-            .requires_reload
-            .map(|value| value.to_string())
-            .unwrap_or_else(|| "-".to_owned())
-    ));
-    lines.push(format!(
-        "filtered_counts by_surface={} by_kind={} reload={} no_reload={}",
-        format_rollup_map(&execution.filtered_action_counts_by_surface),
-        format_rollup_map(&execution.filtered_action_counts_by_kind),
-        execution.filtered_actions_requiring_reload,
-        execution.filtered_actions_without_reload
-    ));
-    lines.extend(render_bridge_profile_fit_lines(&execution.summary));
-    for item in &execution.actions {
-        let remediation_summary = item
-            .supporting_remediations
-            .iter()
-            .map(|support| {
-                let mut parts = vec![support.remediation_class.clone()];
-                if let Some(code) = support.diagnostic_code.as_deref() {
-                    parts.push(format!("code={code}"));
-                }
-                if let Some(field_path) = support.field_path.as_deref() {
-                    parts.push(format!("field={field_path}"));
-                }
-                if support.blocking {
-                    parts.push("blocking=true".to_owned());
-                }
-                parts.join("|")
-            })
-            .collect::<Vec<_>>()
-            .join("; ");
-        lines.push(format!(
-            "- action_id={} surface={} kind={} plugin={} provider={} reload={} follow_up={} supports={} blocked={} warned={} passed={}",
-            item.action.action_id,
-            item.action.surface,
-            item.action.kind,
-            item.action.target_plugin_id,
-            display_text_or_dash(item.action.target_provider_id.as_deref()),
-            item.action.requires_reload,
-            display_text_or_dash(item.action.follow_up_profile.as_deref()),
-            item.supporting_results,
-            item.blocked_results,
-            item.warned_results,
-            item.passed_results
-        ));
-        lines.push(format!(
-            "  source={} manifest={} remediations={}",
-            item.action.target_source_path,
-            display_text_or_dash(item.action.target_manifest_path.as_deref()),
-            remediation_summary
-        ));
-    }
-    lines.join("\n")
-}
-
-#[derive(Debug, Clone)]
-struct ResolvedPluginScanSource {
-    scan_roots: Vec<String>,
-    query: String,
-    limit: usize,
-    bridge_support: Option<ResolvedBridgeSupportSelection>,
-}
-
-impl ResolvedPluginScanSource {
-    fn bridge_support_source(&self) -> Option<String> {
-        self.bridge_support
-            .as_ref()
-            .map(|selection| selection.policy.source.clone())
-    }
-
-    fn bridge_support_sha256(&self) -> Option<String> {
-        self.bridge_support
-            .as_ref()
-            .map(|selection| selection.policy.sha256.clone())
-    }
-
-    fn bridge_support_delta_source(&self) -> Option<String> {
-        self.bridge_support
-            .as_ref()
-            .and_then(|selection| selection.delta_source.clone())
-    }
-
-    fn bridge_support_delta_sha256(&self) -> Option<String> {
-        self.bridge_support.as_ref().and_then(|selection| {
-            selection
-                .delta_artifact
-                .as_ref()
-                .map(|artifact| artifact.sha256.clone())
-        })
-    }
-}
-
-#[derive(Debug, Clone)]
-struct PluginInventoryContext {
-    scan_roots: Vec<String>,
-    query: String,
-    limit: usize,
-    bridge_support_source: Option<String>,
-    bridge_support_sha256: Option<String>,
-    bridge_support_delta_source: Option<String>,
-    bridge_support_delta_sha256: Option<String>,
-    spec: RunnerSpec,
-}
-
-impl PluginInventoryContext {
-    fn bridge_support_provenance(&self) -> Option<PluginsBridgeSupportProvenanceView> {
-        PluginsBridgeSupportProvenanceView::from_fields(
-            self.bridge_support_source.as_deref(),
-            self.bridge_support_sha256.as_deref(),
-            self.bridge_support_delta_source.as_deref(),
-            self.bridge_support_delta_sha256.as_deref(),
-        )
-    }
-}
-
-#[derive(Debug, Clone)]
-struct PluginPreflightContext {
-    scan_roots: Vec<String>,
-    query: String,
-    limit: usize,
-    profile: String,
-    bridge_support_source: Option<String>,
-    bridge_support_sha256: Option<String>,
-    bridge_support_delta_source: Option<String>,
-    bridge_support_delta_sha256: Option<String>,
-    spec: RunnerSpec,
-}
-
-impl PluginPreflightContext {
-    fn bridge_support_provenance(&self) -> Option<PluginsBridgeSupportProvenanceView> {
-        PluginsBridgeSupportProvenanceView::from_fields(
-            self.bridge_support_source.as_deref(),
-            self.bridge_support_sha256.as_deref(),
-            self.bridge_support_delta_source.as_deref(),
-            self.bridge_support_delta_sha256.as_deref(),
-        )
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-struct PluginGovernanceSurfaceContextSpec {
-    pack_id: &'static str,
-    agent_id: &'static str,
-    operator_surface: &'static str,
-    surface_label: &'static str,
-}
-
-fn build_plugin_inventory_context(
-    source: &PluginScanSourceArgs,
-    include_ready: bool,
-    include_blocked: bool,
-    include_deferred: bool,
-    include_examples: bool,
-) -> CliResult<PluginInventoryContext> {
-    let default_limit = default_plugin_inventory_limit();
-    let resolved = resolve_plugin_scan_source(source, default_limit, 100, "plugins inventory")?;
-
-    let mut spec = RunnerSpec::template();
-    spec.pack = VerticalPackManifest {
-        pack_id: "plugin-inventory".to_owned(),
-        domain: "ops".to_owned(),
-        version: "0.1.0".to_owned(),
-        default_route: ExecutionRoute {
-            harness_kind: HarnessKind::EmbeddedPi,
-            adapter: Some("pi-local".to_owned()),
-        },
-        allowed_connectors: BTreeSet::new(),
-        granted_capabilities: BTreeSet::from([Capability::ObserveTelemetry]),
-        metadata: BTreeMap::from([("operator_surface".to_owned(), "plugin_inventory".to_owned())]),
-    };
-    spec.agent_id = "agent-plugin-inventory".to_owned();
-    spec.ttl_s = 120;
-    spec.approval = Some(HumanApprovalSpec {
-        mode: HumanApprovalMode::Disabled,
-        ..HumanApprovalSpec::default()
-    });
-    spec.defaults = None;
-    spec.self_awareness = None;
-    spec.plugin_scan = Some(PluginScanSpec {
-        enabled: true,
-        roots: resolved.scan_roots.clone(),
-    });
-    spec.bridge_support = resolved
-        .bridge_support
-        .as_ref()
-        .map(|selection| selection.policy.profile.clone());
-    spec.bootstrap = None;
-    spec.auto_provision = None;
-    spec.hotfixes = Vec::new();
-    spec.operation = OperationSpec::PluginInventory {
-        query: resolved.query.clone(),
-        limit: resolved.limit,
-        include_ready,
-        include_blocked,
-        include_deferred,
-        include_examples,
-    };
-    let bridge_support_source = resolved.bridge_support_source();
-    let bridge_support_sha256 = resolved.bridge_support_sha256();
-    let bridge_support_delta_source = resolved.bridge_support_delta_source();
-    let bridge_support_delta_sha256 = resolved.bridge_support_delta_sha256();
-
-    Ok(PluginInventoryContext {
-        scan_roots: resolved.scan_roots,
-        query: resolved.query,
-        limit: resolved.limit,
-        bridge_support_source,
-        bridge_support_sha256,
-        bridge_support_delta_source,
-        bridge_support_delta_sha256,
-        spec,
-    })
-}
-
-fn build_plugin_doctor_context(
-    source: &PluginDoctorSourceArgs,
-    include_passed: bool,
-    include_warned: bool,
-    include_blocked: bool,
-    include_deferred: bool,
-) -> CliResult<PluginPreflightContext> {
-    let policy_signature = build_policy_signature_spec(
-        source.policy_signature_algorithm.as_str(),
-        source.policy_signature_public_key_base64.as_deref(),
-        source.policy_signature_base64.as_deref(),
-    )?;
-    let profile = source.profile.as_profile();
-    let surface_spec = PluginGovernanceSurfaceContextSpec {
-        pack_id: "plugin-doctor",
-        agent_id: "agent-plugin-doctor",
-        operator_surface: "plugin_doctor",
-        surface_label: "plugins doctor",
-    };
-
-    build_plugin_preflight_context_from_parts(
-        &source.scan,
-        profile,
-        source.policy_path.clone(),
-        source.policy_sha256.clone(),
-        policy_signature,
-        include_passed,
-        include_warned,
-        include_blocked,
-        include_deferred,
-        false,
-        surface_spec,
-    )
-}
-
-fn build_plugin_preflight_context(
-    source: &PluginGovernanceSourceArgs,
-    include_passed: bool,
-    include_warned: bool,
-    include_blocked: bool,
-    include_deferred: bool,
-    include_examples: bool,
-) -> CliResult<PluginPreflightContext> {
-    let policy_signature = build_policy_signature_spec(
-        source.policy_signature_algorithm.as_str(),
-        source.policy_signature_public_key_base64.as_deref(),
-        source.policy_signature_base64.as_deref(),
-    )?;
-    let profile = source.profile.as_profile();
-    let surface_spec = PluginGovernanceSurfaceContextSpec {
-        pack_id: "plugin-governance",
-        agent_id: "agent-plugin-governance",
-        operator_surface: "plugin_governance",
-        surface_label: "plugins governance",
-    };
-
-    build_plugin_preflight_context_from_parts(
-        &source.scan,
-        profile,
-        source.policy_path.clone(),
-        source.policy_sha256.clone(),
-        policy_signature,
-        include_passed,
-        include_warned,
-        include_blocked,
-        include_deferred,
-        include_examples,
-        surface_spec,
-    )
-}
-
-fn build_plugin_preflight_context_from_parts(
-    scan: &PluginScanSourceArgs,
-    profile: PluginPreflightProfile,
-    policy_path: Option<String>,
-    policy_sha256: Option<String>,
-    policy_signature: Option<SecurityProfileSignatureSpec>,
-    include_passed: bool,
-    include_warned: bool,
-    include_blocked: bool,
-    include_deferred: bool,
-    include_examples: bool,
-    surface_spec: PluginGovernanceSurfaceContextSpec,
-) -> CliResult<PluginPreflightContext> {
-    let default_limit = default_plugin_preflight_limit();
-    let resolved =
-        resolve_plugin_scan_source(scan, default_limit, 500, surface_spec.surface_label)?;
-
-    let mut spec = RunnerSpec::template();
-    spec.pack = VerticalPackManifest {
-        pack_id: surface_spec.pack_id.to_owned(),
-        domain: "ops".to_owned(),
-        version: "0.1.0".to_owned(),
-        default_route: ExecutionRoute {
-            harness_kind: HarnessKind::EmbeddedPi,
-            adapter: Some("pi-local".to_owned()),
-        },
-        allowed_connectors: BTreeSet::new(),
-        granted_capabilities: BTreeSet::from([Capability::ObserveTelemetry]),
-        metadata: BTreeMap::from([(
-            "operator_surface".to_owned(),
-            surface_spec.operator_surface.to_owned(),
-        )]),
-    };
-    spec.agent_id = surface_spec.agent_id.to_owned();
-    spec.ttl_s = 120;
-    spec.approval = Some(HumanApprovalSpec {
-        mode: HumanApprovalMode::Disabled,
-        ..HumanApprovalSpec::default()
-    });
-    spec.defaults = None;
-    spec.self_awareness = None;
-    spec.plugin_scan = Some(PluginScanSpec {
-        enabled: true,
-        roots: resolved.scan_roots.clone(),
-    });
-    spec.bridge_support = resolved
-        .bridge_support
-        .as_ref()
-        .map(|selection| selection.policy.profile.clone());
-    spec.bootstrap = None;
-    spec.auto_provision = None;
-    spec.hotfixes = Vec::new();
-    spec.operation = OperationSpec::PluginPreflight {
-        query: resolved.query.clone(),
-        limit: resolved.limit,
-        profile,
-        policy_path,
-        policy_sha256,
-        policy_signature,
-        include_passed,
-        include_warned,
-        include_blocked,
-        include_deferred,
-        include_examples,
-    };
-    let bridge_support_source = resolved.bridge_support_source();
-    let bridge_support_sha256 = resolved.bridge_support_sha256();
-    let bridge_support_delta_source = resolved.bridge_support_delta_source();
-    let bridge_support_delta_sha256 = resolved.bridge_support_delta_sha256();
-
-    Ok(PluginPreflightContext {
-        scan_roots: resolved.scan_roots,
-        query: resolved.query,
-        limit: resolved.limit,
-        profile: profile.as_str().to_owned(),
-        bridge_support_source,
-        bridge_support_sha256,
-        bridge_support_delta_source,
-        bridge_support_delta_sha256,
-        spec,
-    })
-}
-
-fn resolve_plugin_scan_source(
-    source: &PluginScanSourceArgs,
-    default_limit: usize,
-    max_limit: usize,
-    surface_label: &str,
-) -> CliResult<ResolvedPluginScanSource> {
-    let roots = normalize_scan_roots(&source.roots, surface_label)?;
-    let requested_limit = source.limit.unwrap_or(default_limit);
-    let limit = validate_plugin_limit(requested_limit, max_limit, surface_label)?;
-    let bridge_support = resolve_bridge_support_selection(
-        source.bridge_support.as_deref(),
-        source.bridge_profile.map(PluginBridgeProfileArg::as_str),
-        source.bridge_support_delta.as_deref(),
-        source.bridge_support_sha256.as_deref(),
-        source.bridge_support_delta_sha256.as_deref(),
-    )?;
-
-    Ok(ResolvedPluginScanSource {
-        scan_roots: roots,
-        query: source.query.clone(),
-        limit,
-        bridge_support,
-    })
-}
-
-fn load_bridge_profile_views(
-    requested: &[PluginBridgeProfileArg],
-) -> CliResult<Vec<PluginsBridgeProfileExecutionView>> {
-    let requested = if requested.is_empty() {
-        vec![
-            PluginBridgeProfileArg::NativeBalanced,
-            PluginBridgeProfileArg::OpenclawEcosystemBalanced,
-        ]
-    } else {
-        requested.to_vec()
-    };
-
-    let mut views = Vec::new();
-    let mut seen = BTreeSet::new();
-    for profile in requested {
-        let profile_id = profile.as_str();
-        if !seen.insert(profile_id.to_owned()) {
-            continue;
-        }
-        let resolved =
-            resolve_bridge_support_policy(None, Some(profile_id), None)?.ok_or_else(|| {
-                format!("bundled bridge support profile `{profile_id}` was not resolved")
-            })?;
-        let mut supported_bridges = resolved
-            .profile
-            .supported_bridges
-            .iter()
-            .map(|bridge| bridge.as_str().to_owned())
-            .collect::<Vec<_>>();
-        supported_bridges.sort();
-
-        let mut supported_compatibility_modes = resolved
-            .profile
-            .supported_compatibility_modes
-            .iter()
-            .map(|mode| mode.as_str().to_owned())
-            .collect::<Vec<_>>();
-        supported_compatibility_modes.sort();
-
-        let mut supported_compatibility_shims = resolved
-            .profile
-            .supported_compatibility_shims
-            .iter()
-            .map(|shim| format!("{}:{}", shim.shim_id, shim.family))
-            .collect::<Vec<_>>();
-        supported_compatibility_shims.sort();
-
-        let mut shim_support_profiles = resolved
-            .profile
-            .supported_compatibility_shim_profiles
-            .iter()
-            .map(|profile| {
-                let mut supported_dialects = profile
-                    .supported_dialects
-                    .iter()
-                    .map(|dialect| dialect.as_str().to_owned())
-                    .collect::<Vec<_>>();
-                supported_dialects.sort();
-
-                let mut supported_bridges = profile
-                    .supported_bridges
-                    .iter()
-                    .map(|bridge| bridge.as_str().to_owned())
-                    .collect::<Vec<_>>();
-                supported_bridges.sort();
-
-                let mut supported_adapter_families = profile
-                    .supported_adapter_families
-                    .iter()
-                    .cloned()
-                    .collect::<Vec<_>>();
-                supported_adapter_families.sort();
-
-                let mut supported_source_languages = profile
-                    .supported_source_languages
-                    .iter()
-                    .cloned()
-                    .collect::<Vec<_>>();
-                supported_source_languages.sort();
-
-                PluginsBridgeShimSupportProfileView {
-                    shim_id: profile.shim.shim_id.clone(),
-                    shim_family: profile.shim.family.clone(),
-                    version: profile.version.clone(),
-                    supported_dialects,
-                    supported_bridges,
-                    supported_adapter_families,
-                    supported_source_languages,
-                }
-            })
-            .collect::<Vec<_>>();
-        shim_support_profiles.sort_by(|left, right| {
-            (
-                left.shim_id.as_str(),
-                left.shim_family.as_str(),
-                left.version.as_deref().unwrap_or_default(),
-            )
-                .cmp(&(
-                    right.shim_id.as_str(),
-                    right.shim_family.as_str(),
-                    right.version.as_deref().unwrap_or_default(),
-                ))
-        });
-
-        views.push(PluginsBridgeProfileExecutionView {
-            profile_id: profile_id.to_owned(),
-            source: resolved.source,
-            policy_version: resolved.profile.policy_version.clone(),
-            checksum: resolved.checksum,
-            sha256: resolved.sha256,
-            supported_bridges,
-            supported_compatibility_modes,
-            supported_compatibility_shims,
-            shim_support_profiles,
-            execute_process_stdio: resolved.profile.execute_process_stdio,
-            execute_http_json: resolved.profile.execute_http_json,
-            enforce_supported: resolved.profile.enforce_supported,
-            enforce_execution_success: resolved.profile.enforce_execution_success,
-        });
-    }
-
-    Ok(views)
-}
-
-fn normalize_scan_roots(roots: &[String], surface_label: &str) -> CliResult<Vec<String>> {
-    let mut normalized = Vec::new();
-    let mut seen = BTreeSet::new();
-    for root in roots {
-        let trimmed = root.trim();
-        if trimmed.is_empty() {
-            continue;
-        }
-        if seen.insert(trimmed.to_owned()) {
-            normalized.push(trimmed.to_owned());
-        }
-    }
-    if normalized.is_empty() {
-        return Err(format!(
-            "{surface_label} requires at least one non-empty --root"
-        ));
-    }
-    Ok(normalized)
-}
-
-fn validate_plugin_limit(limit: usize, max_limit: usize, surface_label: &str) -> CliResult<usize> {
-    if !(1..=max_limit).contains(&limit) {
-        return Err(format!(
-            "{surface_label} limit must be between 1 and {max_limit}"
-        ));
-    }
-    Ok(limit)
-}
-
-fn build_policy_signature_spec(
-    algorithm: &str,
-    public_key_base64: Option<&str>,
-    signature_base64: Option<&str>,
-) -> CliResult<Option<SecurityProfileSignatureSpec>> {
-    match (public_key_base64, signature_base64) {
-        (None, None) => Ok(None),
-        (Some(_), None) => {
-            Err("plugins governance policy signature requires --policy-signature-base64".to_owned())
-        }
-        (None, Some(_)) => Err(
-            "plugins governance policy signature requires --policy-signature-public-key-base64"
-                .to_owned(),
-        ),
-        (Some(public_key_base64), Some(signature_base64)) => {
-            Ok(Some(SecurityProfileSignatureSpec {
-                algorithm: algorithm.to_owned(),
-                public_key_base64: public_key_base64.to_owned(),
-                signature_base64: signature_base64.to_owned(),
-            }))
-        }
-    }
-}
-
-fn decode_preflight_bridge_profile_recommendation(
-    report: &SpecRunReport,
-) -> CliResult<Option<PluginPreflightBridgeProfileRecommendation>> {
-    let recommendation_value = report
-        .outcome
-        .get("summary")
-        .and_then(|summary| summary.get("bridge_profile_recommendation"))
-        .cloned()
-        .unwrap_or(Value::Null);
-
-    serde_json::from_value(recommendation_value).map_err(|error| {
-        format!("decode plugin preflight bridge profile recommendation failed: {error}")
-    })
-}
-
-fn decode_plugin_inventory_results(
-    report: &SpecRunReport,
-) -> CliResult<Vec<PluginInventoryResult>> {
-    let results_value = report
-        .outcome
-        .get("results")
-        .cloned()
-        .unwrap_or(Value::Null);
-
-    serde_json::from_value(results_value)
-        .map_err(|error| format!("decode plugin inventory results failed: {error}"))
-}
-
-fn summarize_plugin_inventory_results(
-    results: &[PluginInventoryResult],
-) -> PluginsInventorySummaryView {
-    let returned_plugins = results.len();
-    let mut ready_plugins = 0;
-    let mut setup_incomplete_plugins = 0;
-    let mut blocked_plugins = 0;
-    let mut deferred_plugins = 0;
-    let mut loaded_plugins = 0;
-    let mut source_kind_distribution = BTreeMap::new();
-    let mut bridge_kind_distribution = BTreeMap::new();
-    let mut source_language_distribution = BTreeMap::new();
-    let mut setup_surface_distribution = BTreeMap::new();
-    let mut activation_status_distribution = BTreeMap::new();
-
-    for result in results {
-        let activation_status = result.activation_status.as_deref();
-
-        if activation_status == Some("ready") {
-            ready_plugins += 1;
-        }
-        if activation_status == Some("setup_incomplete") {
-            setup_incomplete_plugins += 1;
-        }
-        if activation_status.is_some_and(plugin_inventory_status_is_blocked) {
-            blocked_plugins += 1;
-        }
-        if result.deferred {
-            deferred_plugins += 1;
-        }
-        if result.loaded {
-            loaded_plugins += 1;
-        }
-
-        increment_rollup_count(&mut source_kind_distribution, result.source_kind.as_str());
-        increment_rollup_count(&mut bridge_kind_distribution, result.bridge_kind.as_str());
-
-        let source_language = result.source_language.as_deref().unwrap_or("unknown");
-        increment_rollup_count(&mut source_language_distribution, source_language);
-
-        let setup_surface = inventory_result_setup_surface_label(result);
-        increment_rollup_count(&mut setup_surface_distribution, setup_surface);
-
-        let status_label = inventory_result_status_label(result);
-        increment_rollup_count(&mut activation_status_distribution, status_label);
-    }
-
-    PluginsInventorySummaryView {
-        returned_plugins,
-        ready_plugins,
-        setup_incomplete_plugins,
-        blocked_plugins,
-        deferred_plugins,
-        loaded_plugins,
-        source_kind_distribution,
-        bridge_kind_distribution,
-        source_language_distribution,
-        setup_surface_distribution,
-        activation_status_distribution,
-    }
-}
-
-fn plugin_inventory_status_is_blocked(status: &str) -> bool {
-    if status == "ready" {
-        return false;
-    }
-
-    if status == "setup_incomplete" {
-        return false;
-    }
-
-    true
-}
-
-fn inventory_result_status_label(result: &PluginInventoryResult) -> &str {
-    let activation_status = result.activation_status.as_deref();
-    let has_activation_status = activation_status.is_some_and(|status| !status.is_empty());
-
-    if has_activation_status {
-        return activation_status.unwrap_or("unknown");
-    }
-
-    if result.deferred {
-        return "deferred";
-    }
-
-    "unknown"
-}
-
-fn inventory_result_setup_surface_label(result: &PluginInventoryResult) -> &str {
-    let setup_surface = result.setup_surface.as_deref();
-    let has_setup_surface = setup_surface.is_some_and(|value| !value.is_empty());
-
-    if has_setup_surface {
-        return setup_surface.unwrap_or("none");
-    }
-
-    let setup_mode = result.setup_mode.as_deref();
-    let has_setup_mode = setup_mode.is_some_and(|value| !value.is_empty());
-
-    if has_setup_mode {
-        return "unspecified";
-    }
-
-    "none"
-}
-
-fn increment_rollup_count(values: &mut BTreeMap<String, usize>, key: &str) {
-    let entry = values.entry(key.to_owned()).or_default();
-    let next_value = entry.saturating_add(1);
-    *entry = next_value;
-}
-
 impl PluginsBridgeSupportProvenanceView {
     fn from_fields(
         source: Option<&str>,
@@ -2636,126 +1375,6 @@ impl PluginsBridgeSupportProvenanceView {
     }
 }
 
-fn decode_preflight_summary(
-    report: &SpecRunReport,
-    bridge_support_provenance: Option<PluginsBridgeSupportProvenanceView>,
-) -> CliResult<PluginsPreflightSummaryView> {
-    let summary_value = report
-        .outcome
-        .get("summary")
-        .cloned()
-        .ok_or_else(|| "decode plugin preflight summary failed: missing summary".to_owned())?;
-    let mut summary: PluginsPreflightSummaryView = serde_json::from_value(summary_value)
-        .map_err(|error| format!("decode plugin preflight summary failed: {error}"))?;
-    summary.bridge_support_provenance = bridge_support_provenance;
-    Ok(summary)
-}
-
-fn decode_preflight_results(report: &SpecRunReport) -> CliResult<Vec<PluginPreflightResult>> {
-    let results_value = report
-        .outcome
-        .get("results")
-        .cloned()
-        .unwrap_or(Value::Null);
-
-    serde_json::from_value(results_value)
-        .map_err(|error| format!("decode plugin preflight results failed: {error}"))
-}
-
-fn summarize_plugin_doctor_results(
-    results: &[PluginPreflightResult],
-    preflight_summary: &PluginsPreflightSummaryView,
-) -> PluginsDoctorSummaryView {
-    let mut activation_ready_plugins: usize = 0;
-    let mut setup_incomplete_plugins: usize = 0;
-    let mut deferred_plugins: usize = 0;
-    let mut loaded_plugins: usize = 0;
-    let mut packages_with_operator_actions: usize = 0;
-    let mut total_recommended_actions: usize = 0;
-    let mut total_operator_actions: usize = 0;
-    let mut bridge_kind_distribution = BTreeMap::new();
-    let mut source_language_distribution = BTreeMap::new();
-    let mut setup_surface_distribution = BTreeMap::new();
-    let mut activation_status_distribution = BTreeMap::new();
-
-    for result in results {
-        let plugin = &result.plugin;
-
-        if result.activation_ready {
-            activation_ready_plugins = activation_ready_plugins.saturating_add(1);
-        }
-
-        if plugin.activation_status.as_deref() == Some("setup_incomplete") {
-            setup_incomplete_plugins = setup_incomplete_plugins.saturating_add(1);
-        }
-
-        if plugin.deferred {
-            deferred_plugins = deferred_plugins.saturating_add(1);
-        }
-
-        if plugin.loaded {
-            loaded_plugins = loaded_plugins.saturating_add(1);
-        }
-
-        let recommended_action_count = result.recommended_actions.len();
-        total_recommended_actions =
-            total_recommended_actions.saturating_add(recommended_action_count);
-
-        let operator_action_count = count_preflight_result_operator_actions(result);
-        total_operator_actions = total_operator_actions.saturating_add(operator_action_count);
-
-        if operator_action_count > 0 {
-            packages_with_operator_actions = packages_with_operator_actions.saturating_add(1);
-        }
-
-        increment_rollup_count(&mut bridge_kind_distribution, plugin.bridge_kind.as_str());
-
-        let source_language = plugin.source_language.as_deref().unwrap_or("unknown");
-        increment_rollup_count(&mut source_language_distribution, source_language);
-
-        let setup_surface = inventory_result_setup_surface_label(plugin);
-        increment_rollup_count(&mut setup_surface_distribution, setup_surface);
-
-        let activation_status = inventory_result_status_label(plugin);
-        increment_rollup_count(&mut activation_status_distribution, activation_status);
-    }
-
-    let packages_requiring_author_attention = preflight_summary
-        .warned_plugins
-        .saturating_add(preflight_summary.blocked_plugins);
-
-    PluginsDoctorSummaryView {
-        matched_plugins: preflight_summary.matched_plugins,
-        returned_plugins: results.len(),
-        passed_plugins: preflight_summary.passed_plugins,
-        warned_plugins: preflight_summary.warned_plugins,
-        blocked_plugins: preflight_summary.blocked_plugins,
-        activation_ready_plugins,
-        setup_incomplete_plugins,
-        deferred_plugins,
-        loaded_plugins,
-        packages_requiring_author_attention,
-        packages_with_operator_actions,
-        total_recommended_actions,
-        total_operator_actions,
-        remediation_counts: preflight_summary.remediation_counts.clone(),
-        bridge_kind_distribution,
-        source_language_distribution,
-        setup_surface_distribution,
-        activation_status_distribution,
-    }
-}
-
-fn count_preflight_result_operator_actions(result: &PluginPreflightResult) -> usize {
-    let mut count = 0_usize;
-    for action in &result.recommended_actions {
-        if action.operator_action.is_some() {
-            count = count.saturating_add(1);
-        }
-    }
-    count
-}
-
 fn action_matches_filters(
     item: &PluginsActionPlanItemView,
     filters: &PluginActionFiltersView,
@@ -2771,196 +1390,14 @@ fn action_matches_filters(
             .is_none_or(|requires_reload| item.action.requires_reload == requires_reload)
 }
 
-fn summarize_filtered_actions(
-    actions: &[PluginsActionPlanItemView],
-) -> (
-    BTreeMap<String, usize>,
-    BTreeMap<String, usize>,
-    usize,
-    usize,
-) {
-    let mut by_surface = BTreeMap::new();
-    let mut by_kind = BTreeMap::new();
-    let mut requiring_reload = 0_usize;
-    let mut without_reload = 0_usize;
-    for item in actions {
-        *by_surface.entry(item.action.surface.clone()).or_default() += 1;
-        *by_kind.entry(item.action.kind.clone()).or_default() += 1;
-        if item.action.requires_reload {
-            requiring_reload = requiring_reload.saturating_add(1);
-        } else {
-            without_reload = without_reload.saturating_add(1);
-        }
-    }
-    (by_surface, by_kind, requiring_reload, without_reload)
-}
-
-fn display_text_or_dash(value: Option<&str>) -> &str {
-    match value {
-        Some(value) if !value.is_empty() => value,
-        _ => "-",
-    }
-}
-
-fn format_csv_or_dash(values: &[String]) -> String {
-    if values.is_empty() {
-        "-".to_owned()
-    } else {
-        values.join(",")
-    }
-}
-
-fn format_rollup_map(values: &BTreeMap<String, usize>) -> String {
-    if values.is_empty() {
-        return "-".to_owned();
-    }
-    values
-        .iter()
-        .map(|(key, value)| format!("{key}:{value}"))
-        .collect::<Vec<_>>()
-        .join(",")
-}
-
-fn write_bridge_support_template(path: &str, template: &BridgeSupportSpec) -> CliResult<()> {
-    let rendered = serde_json::to_string_pretty(template)
-        .map_err(|error| format!("serialize bridge support template failed: {error}"))?;
-    if let Some(parent) = Path::new(path).parent()
-        && !parent.as_os_str().is_empty()
-    {
-        fs::create_dir_all(parent).map_err(|error| {
-            format!(
-                "create bridge template parent directory `{}` failed: {error}",
-                parent.display()
-            )
-        })?;
-    }
-    fs::write(path, rendered)
-        .map_err(|error| format!("write bridge support template `{path}` failed: {error}"))
-}
-
-fn write_bridge_support_delta_artifact(
-    path: &str,
-    artifact: &MaterializedBridgeSupportDeltaArtifact,
-) -> CliResult<()> {
-    let rendered = serde_json::to_string_pretty(artifact)
-        .map_err(|error| format!("serialize bridge support delta artifact failed: {error}"))?;
-    if let Some(parent) = Path::new(path).parent()
-        && !parent.as_os_str().is_empty()
-    {
-        fs::create_dir_all(parent).map_err(|error| {
-            format!(
-                "create bridge delta parent directory `{}` failed: {error}",
-                parent.display()
-            )
-        })?;
-    }
-    fs::write(path, rendered)
-        .map_err(|error| format!("write bridge support delta artifact `{path}` failed: {error}"))
-}
-
-fn render_bridge_profile_fit_lines(summary: &PluginsPreflightSummaryView) -> Vec<String> {
-    let mut lines = vec![format!(
-        "bridge_profiles active={} recommended={} recommended_source={} active_matches={} active_support_fits_all={}",
-        display_text_or_dash(summary.active_bridge_profile.as_deref()),
-        display_text_or_dash(summary.recommended_bridge_profile.as_deref()),
-        display_text_or_dash(summary.recommended_bridge_profile_source.as_deref()),
-        summary
-            .active_bridge_profile_matches_recommended
-            .map(|value| value.to_string())
-            .unwrap_or_else(|| "-".to_owned()),
-        summary
-            .active_bridge_support_fits_all_plugins
-            .map(|value| value.to_string())
-            .unwrap_or_else(|| "-".to_owned())
-    )];
-
-    for fit in &summary.bridge_profile_fits {
-        lines.push(format!(
-            "bridge_profile_fit profile={} version={} fits_all={} supported={} blocked={} reasons={} sample_blocked_plugins={}",
-            fit.profile_id,
-            display_text_or_dash(fit.policy_version.as_deref()),
-            fit.fits_all_plugins,
-            fit.supported_plugins,
-            fit.blocked_plugins,
-            format_rollup_map(&fit.blocking_reasons),
-            format_csv_or_dash(&fit.sample_blocked_plugins)
-        ));
-    }
-
-    if let Some(recommendation) = summary.bridge_profile_recommendation.as_ref() {
-        lines.push(format!(
-            "bridge_profile_recommendation kind={} target={} source={} version={} summary={}",
-            recommendation.kind,
-            recommendation.target_profile_id,
-            recommendation.target_profile_source,
-            display_text_or_dash(recommendation.target_policy_version.as_deref()),
-            recommendation.summary
-        ));
-        if let Some(delta) = recommendation.delta.as_ref() {
-            lines.push(format!(
-                "bridge_profile_delta bridges={} compatibility={} adapter_families={} shims={} shim_profiles={} unresolved={}",
-                format_csv_or_dash(&delta.supported_bridges),
-                format_csv_or_dash(&delta.supported_compatibility_modes),
-                format_csv_or_dash(&delta.supported_adapter_families),
-                format_csv_or_dash(&delta.supported_compatibility_shims),
-                format_shim_profile_deltas(&delta.shim_profile_additions),
-                format_csv_or_dash(&delta.unresolved_blocking_reasons)
-            ));
-        }
-    }
-
-    lines
-}
-
-fn format_shim_profile_deltas(values: &[PluginsBridgeShimProfileDeltaView]) -> String {
-    if values.is_empty() {
-        return "-".to_owned();
-    }
-
-    values
-        .iter()
-        .map(|value| {
-            format!(
-                "{}:{}:dialects={}|bridges={}|adapter_families={}|languages={}",
-                value.shim_id,
-                value.shim_family,
-                format_csv_or_dash(&value.supported_dialects),
-                format_csv_or_dash(&value.supported_bridges),
-                format_csv_or_dash(&value.supported_adapter_families),
-                format_csv_or_dash(&value.supported_source_languages)
-            )
-        })
-        .collect::<Vec<_>>()
-        .join(";")
-}
-
-fn format_bridge_shim_profile_delta_artifact(
-    values: &[crate::PluginPreflightBridgeShimProfileDelta],
-) -> String {
-    if values.is_empty() {
-        return "-".to_owned();
-    }
-
-    values
-        .iter()
-        .map(|value| {
-            format!(
-                "{}:{}:dialects={}|bridges={}|adapter_families={}|languages={}",
-                value.shim_id,
-                value.shim_family,
-                format_csv_or_dash(&value.supported_dialects),
-                format_csv_or_dash(&value.supported_bridges),
-                format_csv_or_dash(&value.supported_adapter_families),
-                format_csv_or_dash(&value.supported_source_languages)
-            )
-        })
-        .collect::<Vec<_>>()
-        .join(";")
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::default_plugin_inventory_limit;
+    use crate::plugins_cli::governance_support::{
+        build_policy_signature_spec, normalize_scan_roots,
+    };
+    use crate::plugins_cli::render_support::wrap_plugins_surface_text;
     use crate::{
         PLUGIN_PREFLIGHT_SUMMARY_SCHEMA_PURPOSE, PLUGIN_PREFLIGHT_SUMMARY_SCHEMA_SURFACE,
         PLUGIN_PREFLIGHT_SUMMARY_SCHEMA_VERSION,
@@ -2976,6 +1413,10 @@ mod tests {
             .join(format!("{prefix}-{nanos}"))
             .display()
             .to_string()
+    }
+
+    fn wrap_plugins_surface_text(title: &str, body: String) -> String {
+        crate::render_operator_shell_surface_from_body(title, "operator plugins", body)
     }
 
     #[test]
